+++ public/app/partials/systems.html
@@ -52,26 +52,26 @@
                 <div style="width:70%; margin:0 auto;">
                     <div class="system-status-box pull-left">
                         <span class="system-label system-label-success">正常</span>
-                        <span class="system-legend system-legend-success">{{seriesStatus.normal}}</span>
+                        <span class="system-legend system-legend-success">{{serviesStatus.normal}}</span>
                     </div>
                     <div class="system-status-box pull-right">
                         <span class="system-label system-label-warning">异常</span>
-                        <span class="system-legend system-legend-warning">{{seriesStatus.unnormal}}</span>
+                        <span class="system-legend system-legend-warning">{{serviesStatus.unnormal}}</span>
                     </div>
                     <div class="clearfix"></div>
                     <div ng-if="servies.length" class="system-status-percent">
-                        <span style="width: {{seriesStatus.normal / servies.length * 100}}%;">{{seriesStatus.normal / servies.length * 100 |limitTo:4}}%</span>
-                        <span style="width: {{seriesStatus.unnormal / servies.length * 100}}%;">{{seriesStatus.unnormal / servies.length * 100 |limitTo:4}}%</span>
+                        <span style="width: {{serviesStatus.normal / servies.length * 100}}%;">{{serviesStatus.normal / servies.length * 100 |limitTo:4}}%</span>
+                        <span style="width: {{serviesStatus.unnormal / servies.length * 100}}%;">{{serviesStatus.unnormal / servies.length * 100 |limitTo:4}}%</span>
                     </div>
                     <div ng-if="!servies.length" class="system-status-percent">
                         暂无服务状态
                     </div>
                     <div class="status-list">
-                        <div class="status-normal" style="width: {{seriesStatus.normal / servies.length * 100}}%;">
-                            <span ng-repeat="status in servies | filter:{'status': 0}">{{status.name}}</span>
+                        <div class="status-normal" style="width: {{serviesStatus.normal / servies.length * 100}}%;">
+                            <span ng-repeat="status in servies | filter:{'alive': 0}">{{status.service}}</span>
                         </div>
-                        <div class="status-unnormal" style="width: {{seriesStatus.unnormal / servies.length * 100}}%;">
-                            <span ng-repeat="status in servies | filter:{'status': 1}">{{status.name}}</span>
+                        <div class="status-unnormal" style="width: {{serviesStatus.unnormal / servies.length * 100}}%;">
+                            <span ng-repeat="status in servies" ng-if="status.alive !== 0">{{status.service}}</span>
                         </div>
                     </div>
                     <div ng-if="servies.length">查看详情</div>
@@ -100,10 +100,10 @@
                     </div>
                     <div class="status-list">
                         <div class="status-normal"  style="width: {{hostStatus.normal / hostList.length * 100}}%;">
-                            <span ng-repeat="status in hostList | filter:{'status': 0}">{{status.host}}</span>
+                            <span ng-repeat="status in hostList | filter:{'collector.state': 0}">{{status.hostname}}</span>
                         </div>
                         <div class="status-unnormal" style="width: {{hostStatus.unnormal / hostList.length * 100}}%;">
-                            <span ng-repeat="status in hostList | filter:{'status': 1}">{{status.host}}</span>
+                            <span ng-repeat="status in hostList" ng-if="status['collector.state'] !== 0">{{status.hostname}}</span>
                         </div>
                     </div>
                 </div>
